Add Symbian-specific implementation of QGraphicsVideoItem
This implementation creates a separate top-level widget for each
QGraphicsVideoItem.  This widget is positioned behind the viewport,
and a transparent hole is painted in the graphics view.

The ordinal position of the native window owned by the top-level
widget is set such that the stacking order of these windows
matches that of the corresponding QGraphicsVideoItems.

Task-number: QTMOBILITY-727
Reviewed-by: Derick Hawcroft
Reviewed-by: Dmytro Poplavskiy