Stops Location code from crashing when capabilities are missing.
Symbian programs that use the Location API for positioning must
have the Location capability.

Prior to this change the applications would crash if the
capability was missing.  This change prevents the crash, instead
logging a warning and returning 0 for the position source,
satellite source or area monitor, which indicates that there is
no platform support.