Fixes map viewer example.
Several signals where still being used which don't exist anymore as
their implementation has been triaged to the next release.

Unaccepted events weren't being passed through to the Declarative
items beneath them.  This was fixed, and the example was modified
to put the map above the MouseArea so the mouse based navigation
could be used in conjunction with the MapMouseArea code.

This is just a work around until we have proper support for mouse
movement in the QML bindings for maps.