Fixed possible null-pointer dereference when QAudioInput is closed
In push mode, where there is additional data available, a method call
is queued. If QAudioInput is closed before this call is invoked, then
the pulse audio plugin will crash, because the queued method call tries
to dereference the audio source which has been deleted and set to zero
in the close method.

Task-number: QTMOBILITY-1367
Reviewed-by: Michael Goddard