Increment m_totalTimeValue also when reading from the temp buffer
Since m_totalTimeValue was not updated when data was written to the
client from the temp buffer, then processedUSecs() would become more and
more inaccurate if client provided a buffer for QAudioInput that was too
small.

Changed processedUSecs to use pa_bytes_to_usec rather than calculating
the time from the bytes ourselves.

Updated use of buffer size in audioinput example to make it easier to
change in testing.

Task-number: QTMOBILITY-1368
Reviewed-by: Derick Hawcroft