Fix failing auto tests for symbian 5th ed
Also fix a bug that was found during testing where MatchContains
was being falsely detected when MatchEnd was actually being used
This was due to the fact that MatchContains is 0x2 while
MatchEnd is 0x3

Also in fetchWaitForFinished() limit the number of landmarks imported.
There seems to be a race condition bug when multiple imports of
large numbers of landmarks are performed.  This issue has
been raised in MOBILITY-2664.

Task-number: MOBILITY-2663