Fixes various unit tests.
The QGeoCoordinate unit test was failing on platforms which
use float for qreal.

One of the tiled mapping helper class unit tests was failing because
it appears to have been written to test for the existing (rather than
the correct) behaviour.  Both the behaviour and the test have been
fixed.