Use BlitWriteAlpha in QGraphicsVideoItem on Symbian
In order for the direct camera viewfinder to be visible on DSA devices,
alpha=0 must be written into the framebuffer in the region where the
viewfinder will be displayed.  This requires the backing store to have
an alpha channel, and for the blitting of the backing store to use
CGraphicsContext::EDrawModeWriteAlpha (in order to propagate alpha values
directly into the frame buffer rather than blending).

This patch uses private QtGui APIs to achieve the required behaviour.

Note that this requires changes in QtGui; the relevant changes in the
Qt repository are tagged QTMOBILITY-1278.

Task-number: QTMOBILITY-1278
Task-number: QTMOBILITY-1279
Task-number: QTMOBILITY-1334
Reviewed-by: Ruth Sadler
Reviewed-by: Tuomas Viitanen