Bug id: ou1cimx1#740266
Case: Profiles: Profile cannot be edited in Profile application.
Root cause:
If there isn't the default mass storage in the phone, the image directory can't be created and cause cntsrv.exe panic in the case.
Because cntsrv.exe isn't work in the case, contacts model can't support service.

solution:
If there isn't the default mass storage in the phone,or the path is invalid in the default mass storage, the leave from TCntImageRescaleUtility::CreateImageDirectoryL should be handled in CPplContactsFile.

Reviewed-by Ilya Sinitsyn