Initialise specialised NDEF record types.
Depending on the record type specifications a record with an empty
payload might not be a valid record. For example the Text record has a
minium size of 1 byte (the status byte), the locale and text may be of
zero size as specified by the status byte and the payload length. Uri
records also have a 1 byte minium size (the uri prefix).