QuickTime backend of QtMultimedia reports QMediaPlayer::StalledMedia way too early.
kMovieLoadStateLoading has no clear distinction of loading like
the first load and loading to get more data because the playback can't continue.
The code in QtMM assumed that whenever it's not stopped and we reach
kMovieLoadStateLoading, it is stalled. Unfortunately this is not true.
Because if we sent the stalled state to the users it means the media has
been loaded and therefore we can get the meta-data for example. In that
particular use case it's wrong, it's the first load so the backend lie
by saying a state way more ahead of what it actually have.
The proposed patch ensure that in order to send the QMediaPlayer::StalledMedia
we reached previously at least the QMediaPlayer::LoadedState.

Task-number:QTMOBILITY-1463
Reviewed-by:Dmytro Poplavskiy

Merge-request: 28
Reviewed-by: Michael Goddard <michael.goddard@nokia.com>