Fix up the error handling when the manager calls the engines.
When the engine is called by the manager, it was passed the manager's
error state directly, so that it could be modified.  But if the engine
emitted signals, and slots connected to those themselves called back
into the manager, the error state could be overwritten.  This is
particularly noticeable when the engine does something like
"return *error == QContactManager::NoError;" after emitting signals, on
the assumption that the error state was still valid.

Change-Id: Iaa17ab6471fd91c22f46760e5ae5974a6c0fb487
Task-number: QTMOBILITY-1499
Reviewed-by: Chris Adams
Reviewed-by: Xizhi Zhu