Initial creation of the Windows backend for camera
This implementation uses DirectShow to grab single images which are then transfered to a surface. This is not the optimal solution for the platform, however, the Qt Widget it is placed in is not guaranteed to have a native Windows handle, and this would have been a requirement to use the better option.

Reviewed-by: Ling Hu