Provide default import file for declarative landamrk example
Previously the import file was not found if the landmarkmap example
was started from a different directory to that which the binary
was located.  This fix is a workaround for that.

Ideally what should be happening is that we use resource files
but since this does not work with symbian yet we need to workaround it.

Task-number: MOBILITY-2735