Refuse to build any example/demos on Freemantle
Mobility 1.2 no longer supports Freemantle/Maemo5. Since new symbols
have been added to the libraries the examples and demos are not
guaranteed to even build on Ffreemantle anymore. Some may still build
but it's up to the developer wanting to try the example to establish
this.

The most appropriate way of disabling Maemo5 support would be to remove
Maemo5 as a target from the entire source package. This takes
significantly more time and is likely to be done later on. Right
now this patch ensures that Qt SDK users wanting to compile examples
are notified that they cannot compile the examples for Freemantle.

Task-number: QTSDK-670