make it easier to assign a category to a landmark
Minor fix so that clicking the checkbox itself will also cause
the checkbox to toggle.

Previously the categoryPressed slot would cause a toggle, but so
would the checkbox itsel, when the checkbox is pressed
causing them to cancel out.  Therefore make the item non user
checkable and just let the categoryPressed slot alone handle
the toggle.

Task-number: QTMOBILITY-1564