Fix bug QTMOBILITY-1058 in SystemInfo Symbian backend.
On Symbian instantiate a separate private object for each instance of the public
class instead of using a global static singleton. This guarantees that the
private class and all related Symbian backend objects are accessed and
finally destroyed in the thread where they were created.

The private singleton is still used in all static methods. In addition, this
implementation does not support moveToThread() on Symbian.

Task-number: QTMOBILITY-1058
Reviewed-by: Iiro Kause