Partial fix for organizer mkcal backend async API bug
When starting the processing of an async request, the request state was
changed to ActiveState only after starting the worker thread responsible for
carrying out the request, see OrganizerAsynchManager::startRequest. In many
cases the request had already been completed when the request state was
updated, and update changed state from FinishedState to ActiveState.

The state update was moved to happen before starting the worker thread.

Change-Id: I474d7074b46b8a9111c747a005d663939029c14d
Task-number: MOBILITY-2731