Fixes several Maps misbehaviours
There were a few problems here.

The viewport calcuation for the tiled mapping code was incorrect when
the map was zoomed out to the point that the map is smaller than the
window that its contained in.  This caused objects to disappear
at small zoom levels.

Some tests were failing because they assumed qreal was a double.

Some tests were failing because mapObjectsAtScreenPosition was
implemented as if it was mapObjectsInScreenRect for a 3x3 rectangle
centered on the point, but with a different and less robust
implementation.  The implementation is now call to
mapObjectsInScreenRect.

Some test were failing because the behaviour of offset text,
pixmap and custom objects was incorrect.  This was corrected
and the associated testing was increased in order to reduce
the chance we run into that again.