Gst player backend: fixed initial buffering
If buffer is not full it's not enough to check the current session state
for playing state to pause but pending state as well.

Otherwise since the playing state is not reached yet pause() is not
called and playback starts for a short period of time as soon
as pipeline is prerolled untill next buffering message is received.

Task-number: PMO 254018
Reviewed-by: Michael Goddard