Fixes incorrect bounding box behaviour.
This fixes some incorrect behaviour with the bounding box union
operations.

A subset of the failing use cases - where the bounding boxes are
180 degrees apart - are best addressed by returning a bounding
box 360 degrees wide so that result is independent of the order
of operations.

This is "new" behaviour, but since the "old" behaviour was
masked by this bug and this bug has existed since the
initial release it seems like now is the best time to
change this behaviour.

The unit tests and documentation were also updated.