Bug id: ou1cimx1#756589
Case Title: Genius case #177766:7-00: Customer have added 5300 business cards to his E7-00, after call does not succeed.
Root cause:
Actually the case works as expected that Phonebook is announcing that memory is full (and asks to close some applications before continuing) and is
still up and running. Phonebook is mainly able to complete small tasks without announcing that memory is low, but if doing big tasks as
searching in find box, sometimes it will announce that memory is full if the amount of matched contacts is huge.

cntsrv (cntsrv.exe) was memory low for this case, and the reason is that when searching contacts in find box, cntsrv allocates memory for every
matched contact (CViewContact::NewL) and append it to an array, so the more matched contacts we have, the more heap memory is needed in cntsrv.

Solution:
Increase cntsrv heap size from 2MB to 8MB to fix the case for SR1.11, and further optimization for cntsrv RAM consumption will be done by Espoo through a separate task.
After this change at least 10000 contacts could be supported without poping up memory low notification.

Reviewed-by Ilya Sinitsyn