Fix some threading issues with QSampleCache.
It was possible to be in QSample::release() after the dtor is called
from the loading thread.  Also, QSampleCache should be marked internal
and doesn't need to inherit from QObject.

Rearrange the locking a little bit, and store samples that are in
the process of being deleted in a separate set, so when the loading
thread is killed before they get deleted, we can still clean them up.

Reviewed-by: Jonas Rabbe