Connect signals lazily with the V2 engine wrapper too.
Previously the QContactManager would lazily connect to the engine signals
when the manager was connected to, but unfortunately the V2 engine
wrapper itself was always connecting to the wrappee's signals.  So
V2 engines got lazy signals, but V1 engines (most of them) did not.  So
now tweak the signal connection code a bit to take advantage of the
simple signal proxying that the wrapper used to do.  This might need
refactoring in the future if a V3 wrapper emits signals itselfs, though.

Change-Id: Id75e40a8da09133a53cef9ce18ec0d5e430e7fbc
Task-number: QTMOBILITY-507
Reviewed-by: Chris Adams