Fixed directshow config test
In some versions of Microsoft Windows SDK v6.0A, the qedit.h header is
present, but dxtrans.h (which is #included by qedit.h) is missing.

The standard workaround for this issue is already used by
plugins/multimedia/directshow/camera/dscamerasession.h; this commit adds
the same workaround to the config test.

Reviewed-by: Michael Goddard
Reviewed-by: Ling Hu