SymbianCameraBE: Restructure advanced camera settings
Remove the dependency to the S60CameraAdvSettings from the control classes.
This makes the code safer as the advanced feature settings happens
always via S60ImageCaptureSession and existence of the
S60CameraAdvSettings is not needed to be promoted for the control classes.
Also, add support for setting the advcanced camera settings in the
UnloadedStatus. This can be achieved by saving the requested value if
camera is not loaded and applying the requested values when loading
the camera.

Task-number: QTMOBILITY-1566
Reviewed-by: Derick Hawcroft
Change-Id: I494990baaf134ae8adc2832722a9d1156b72f136