Destroy media player session when no longer required
Previously, the mediaplayer backend lazily constructed, and then kept in
existence, an S60AudioPlayerSession and an S60VideoPlayerSession instance.
When either of these was no longer required (i.e. media type switched from
audio to video, video to audio, or to null), the
S60MediaPlayerSession::reset() function was used to return the player
session to its base state.

This function does not properly clean up, however, meaning that in some
circumstances (e.g. video resources being lost), subsequent attempts
to use the session fail silently.

This patch replaces the previous mechanism with one in which
S60MediaPlayerSession object lifetime is directly tied to the period during
which the session is in use.

Note that this patch includes moving the emission of QMediaControl::error(...)
to the end of S60MediaPlayerSession::setError.  This is because, if the
client decides, on receiving the error, to set another source, the current
session will be destroyed.

Reviewed-by: Dmytro Poplavskiy