Resync with current qt-mobility master
Now this harmattan-pr1.2 branch is at the same state as
qt-mobility master at commit id
c47f1b3480c78739615c868f958b138bfafa67b8 with the difference
that the src/harmattaninstalls/ directory is copied over to
debian/ so that we can easily build debian packages from
this repository and branch.

All the patches that were in this branch were added as patches
in qt-mobility master in the src/harmattaninstalls/patches
directory. that way we do have an easier overview about which
patches we do have to maintain on top of qt-mobility master.