Remove Symbian-specific capability check from QDeclarativeCamera
The camera backend now gracefully handles failure to initialize the
camera due to missing capabilities (see commit 99a3f630).
Therefore it is no longer necessary to have a check in the
QDeclarativeCamera constructor.

This means that we no longer dereference a null m_viewfinderItem in
QDeclarativeCamera::geometryChanged(), which was the cause of this
crash.

Task-number: QTMOBILITY-1438
Reviewed-by: Jonas Rabbe