Added qmlvideo and qmlvideofx demos
qmlvideo allows the user to select from a menu one of a number of
video or camera use cases.  These include scenes in which the Video
or Camera item is moved, resized, rotated etc.  This application
is intended to test the various manipulations which can be applied
to Video and Camera items.

qmlvideofx uses the ShaderEffectItem element to apply post-processing
effects, expressed in GLSL, to a Video or Camera item.

While the main body of each app is implemented in QML, each one also
makes use of native code in order to gather and calculate performance
metrics, namely the QML frame rate and (on Symbian only) the graphics
memory consumption.  These metrics are displayed in QML elements
which are overlaid on the UI.

qmlvideofx also demonstrates the use of different top-level QML files
to handle different screen sizes.  On small-screen devices, menus are
by default hidden, and only appear when summoned by a gesture.
Large-screen devices show a more traditional layout in which menus
are displayed around the video content pane.

Reviewed-by: Jonas Rabbe