Fix for adding details in qmlcontacts demo
Fixes a problem when adding details to a contact. The Repeater used in the UI uses list<string> as a model.
These do not update automatically, but re-assigning the model helps.
Furthermore, the phoneNumbers and emails list is not created until the contact is saved (when creating them dynamically).
This is worked around by using a base contact component.

Reviewed-by: Mikko Suonio
Task-number: QTMOBILITY-1514