Fix sending files through QBluetoothTransferManager on Symbian
The problem was in QBluetoothTransferReplySymbian. That class does
first a service discovery for ObexObjectPush service, and then sends
the object through the service. The service discovery finished signal
was wrongly connected, which caused the QBluetoothTransferReplySymbian
object to be deleted right after the service discovery, thus cancelling
the file transfer.
  This change makes sure that QBluetoothServiceDiscoveryAgent is deleted
correctly in all situations. This also fixes problems when the transfer
is cancelled during the service discovery phase.

Task-number: ou1cimx1#957795
Reviewed-by: Mika Laine