Skip some tst_QGraphicsVideoItem steps on Symbian
Several of the steps in tst_QGraphicsVideoItem assume that
QGraphicsVideoItem renders via QVideoRendererControl, which is not the
case for some implementations.  On versions of Symbian which do not have
the EGL_NOK_image_endpoint2 extension for example, video rendering is
done via the QVideoWidgetControl interface.

For most of these steps, only certain assertions are based on this
assumption; after removing these assertions, the test steps can still be
run on platforms which do not use QVideoRendererControl, and can still
perform useful checks.

For the show() and boundingRect() test steps however, the logic of the
test step requires QVideoRendererControl to be used, and therefore these
steps must be skipped on platforms which do not use that interface.

Reviewed-by: mread