Fix tst_QGraphicsVideoItem_symbian
On versions of the platform which lack the EGL_NOK_image_endpoint2
extension, video rendering is always done via the 'direct' path
(i.e. QVideoWidgetControl, rather than QVideoRendererControl).  This
patch adds a runtime check for the presence of that extension, and
adjusts the expected rendering path used in subsequent checks.

Reviewed-by: mread