Fix a crash in S60MediaPlayerControl::setMedia
When S60MediaPlayerControl::setMedia is called, a new
S60MediaPlayerSession is created and the old one is deleted. This can
cause problems if setMedia is called as a result of a signal from the
session. Change the deletion of old session to deleteLater() to avoid
this problem.

Task-number: ou1cimx1#993581
Reviewed-by: Juha Kukkonen