Introduce a QOrientableSensorBase class.
All axes-based sensor and the compass now inherit from this class.
It allows adapting the reading values based on the current screen
orientation.

AxesOrientationMode enum must be replicated in all child classes
for it to work in QML.

Unfortunetely the blackberry NDK was released to public with
this BIC change so the only way to upstream the blackberry
sensors plugins in a BC manner is to #ifdef it out on
!qnx platforms.

Change-Id: I3c49fcc5a116a9ce4f314a95097be637a2e88567
Reviewed-by: Thomas McGuire <thomas.mcguire@kdab.com>
Reviewed-by: Lorn Potter <lorn.potter@gmail.com>