Fix query syntax for PhotoAlbum children.
The nie:isLogicalPartOf(?x) syntax won't match against an id, so this
query would never return anything.  Instead return the contents
of the list rather than relying on a connection that may never
be established.

The same query will also return the contents of a Playlist, support
for which was mysteriously absent.

Change-Id: Id448e06f8d183e064c8966b22d44414057a45f8a
Reviewed-by: Christopher Adams <chris.adams@jollamobile.com>