Use real properties for QSensor::bufferSize & co
The #ifdef Q_DOC hack is not really needed and was
confusing.

In addition, expose those properties to the QML API.

As a result, the backends can now listen to the
bufferSizeChanged() signal to update the buffering
while the sensor is running. This has been implemented
for the BlackBerry platform.

This is a backport of
QtSensors commit 4d9edb754b3dff3331c8d8790a04548f0981f315

Change-Id: Ia34fc1ad2e59552355f1e7d27a67000c29a9900f
Reviewed-by: Bernd Weimer <bweimer@rim.com>
Reviewed-by: Lorn Potter <lorn.potter@jollamobile.com>